requirements-docker.txt
# requirements-docker.txt
# Core dependencies
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.23.0
scipy>=1.10.0
pandas>=2.0.0
scikit-learn>=1.2.0

# Deep Learning
transformers>=4.35.0
accelerate>=0.24.0
datasets>=2.14.0
tokenizers>=0.14.0

# LLM APIs
openai>=1.3.0
anthropic>=0.7.0
google-generativeai>=0.3.0

# RL and MAS
gymnasium>=0.29.0
stable-baselines3>=2.1.0
ray[tune]>=2.8.0

# Web automation
selenium>=4.15.0
playwright>=1.40.0
beautifulsoup4>=4.12.0
requests>=2.31.0

# API and serving
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.5.0
python-multipart>=0.0.6

# Monitoring and logging
wandb>=0.16.0
tensorboard>=2.15.0
prometheus-client>=0.19.0
grafana-api>=1.0.3

# Infrastructure
redis>=5.0.0
celery>=5.3.0
docker>=7.0.0
kubernetes>=28.1.0

# Utilities
python-dotenv>=1.0.0
click>=8.1.0
rich>=13.7.0
tqdm>=4.66.0
pyyaml>=6.0.1
networkx>=3.2.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0

# RunPod SDK
runpod>=1.4.0